enable sending data via post in ajax
