+++ api/src/main/java/org/openmrs/module/ugandaemrsync/tasks/SendReportsTask.java
@@ -4,25 +4,11 @@ import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.apache.http.HttpResponse;
 import org.apache.http.HttpStatus;
-import org.openmrs.api.context.Context;
-import org.openmrs.module.reporting.evaluation.EvaluationContext;
-import org.openmrs.module.reporting.evaluation.parameter.Mapped;
-import org.openmrs.module.reporting.report.ReportData;
-import org.openmrs.module.reporting.report.ReportDesign;
-import org.openmrs.module.reporting.report.ReportRequest;
-import org.openmrs.module.reporting.report.definition.ReportDefinition;
-import org.openmrs.module.reporting.report.definition.service.ReportDefinitionService;
-import org.openmrs.module.reporting.report.renderer.RenderingMode;
-import org.openmrs.module.reporting.report.renderer.TextTemplateRenderer;
-import org.openmrs.module.reporting.report.service.ReportService;
 import org.openmrs.module.ugandaemrsync.api.UgandaEMRHttpURLConnection;
 import org.openmrs.module.ugandaemrsync.server.SyncGlobalProperties;
 import org.openmrs.scheduler.tasks.AbstractTask;
-import org.openmrs.util.OpenmrsUtil;
+import org.openmrs.ui.framework.SimpleObject;
 import org.springframework.stereotype.Component;
-
-import java.io.*;
-import java.util.*;
 
 import static org.apache.commons.lang.StringUtils.isBlank;
 import static org.openmrs.module.ugandaemrsync.UgandaEMRSyncConfig.*;
@@ -35,25 +21,19 @@ import static org.openmrs.module.ugandaemrsync.UgandaEMRSyncConfig.*;
 public class SendReportsTask extends AbstractTask {
 
     protected Log log = LogFactory.getLog(getClass());
-    boolean sent=false;
+     boolean sent=false;
 
-     Date startDate;
+     String previewBody;
 
-     Date endDate;
-
-     String reportDefinitionUuid;
+     SimpleObject response;
 
     public SendReportsTask() {
         super();
     }
 
-    public SendReportsTask(Date startDate, Date endDate, String reportDefinitionUuid) {
-        this.startDate = startDate;
-        this.endDate = endDate;
-        this.reportDefinitionUuid = reportDefinitionUuid;
+    public SendReportsTask(String previewBody) {
+        this.previewBody = previewBody;
     }
-
-
 
     UgandaEMRHttpURLConnection ugandaEMRHttpURLConnection = new UgandaEMRHttpURLConnection();
 
@@ -85,53 +65,19 @@ public class SendReportsTask extends AbstractTask {
             return;
         }
         log.info("Sending Report to server ");
-        String bodyText = generateReport(reportDefinitionUuid,startDate,endDate);
 
-        HttpResponse httpResponse = ugandaEMRHttpURLConnection.httpPost(reportsServerUrlEndPoint, bodyText, syncGlobalProperties.getGlobalProperty(GP_DHIS2_ORGANIZATION_UUID), syncGlobalProperties.getGlobalProperty(GP_DHIS2_ORGANIZATION_UUID));
-        if (httpResponse.getStatusLine().getStatusCode() == HttpStatus.SC_OK || httpResponse.getStatusLine().getStatusCode() == HttpStatus.SC_CREATED) {
-            sent=true;
-            log.info("Report  has been sent to central server");
-        } else {
-            log.info("Http response status code: " + httpResponse.getStatusLine().getStatusCode() + ". Reason: " + httpResponse.getStatusLine().getReasonPhrase()); }
-
-    }
-
-    private String generateReport(String uuid, Date startDate, Date endDate){
-        String strOutput="";
-        try{
-           ReportDefinition rd= getReportDefinitionService().getDefinitionByUuid(uuid);
-            if (rd == null) {
-                throw new IllegalArgumentException("unable to find  report with uuid "
-                        + uuid);
-            }else {
-                List<ReportDesign> reportDesigns = Context.getService(ReportService.class).getReportDesigns(rd, TextTemplateRenderer.class, false);
-                ReportDesign reportDesign = reportDesigns.stream().filter(p -> "JSON".equals(p.getName())).findAny().orElse(null);
-                String reportRendergingMode = JSON_REPORT_RENDERER_TYPE + "!" + reportDesign.getUuid();
-                RenderingMode renderingMode = new RenderingMode(reportRendergingMode);
-                if (!renderingMode.getRenderer().canRender(rd)) {
-                    throw new IllegalArgumentException("Unable to render Report with " + reportRendergingMode);
-                }
-                Map<String, Object> parameterValues = new HashMap<String, Object>();
-
-                parameterValues.put("endDate", endDate);
-                parameterValues.put("startDate", startDate);
-                EvaluationContext context = new EvaluationContext();
-                context.setParameterValues(parameterValues);
-                ReportData reportData = getReportDefinitionService().evaluate(rd, context);
-                ReportRequest reportRequest = new ReportRequest();
-                reportRequest.setReportDefinition(new Mapped<ReportDefinition>(rd, context.getParameterValues()));
-                reportRequest.setRenderingMode(renderingMode);
-                File file = new File(OpenmrsUtil.getApplicationDataDirectory() + "sendReports");
-                FileOutputStream fileOutputStream = new FileOutputStream(file);
-                renderingMode.getRenderer().render(reportData, renderingMode.getArgument(), fileOutputStream);
-
-                strOutput = readOutputFile(strOutput);
+        if(previewBody!="") {
+            HttpResponse httpResponse = ugandaEMRHttpURLConnection.httpPost(reportsServerUrlEndPoint, previewBody, syncGlobalProperties.getGlobalProperty(GP_DHIS2_ORGANIZATION_UUID), syncGlobalProperties.getGlobalProperty(GP_DHIS2_ORGANIZATION_UUID));
+            if (httpResponse.getStatusLine().getStatusCode() == HttpStatus.SC_OK || httpResponse.getStatusLine().getStatusCode() == HttpStatus.SC_CREATED) {
+                sent = true;
+                log.info("Report  has been sent to central server");
+            } else {
+                log.info("Http response status code: " + httpResponse.getStatusLine().getStatusCode() + ". Reason: " + httpResponse.getStatusLine().getReasonPhrase());
             }
-        }catch (Exception e){
-            e.printStackTrace();
         }
-        return strOutput;
     }
+
+
 
 
     public boolean isGpReportsServerUrlSet() {
@@ -150,33 +96,14 @@ public class SendReportsTask extends AbstractTask {
         this.sent = sent;
     }
 
-    public Date getStartDate() {
-        return startDate;
+    public String getPreviewBody() {
+        return previewBody;
     }
 
-    public void setStartDate(Date startDate) {
-        this.startDate = startDate;
+    public void setPreviewBody(String previewBody) {
+        this.previewBody = previewBody;
     }
 
-    public Date getEndDate() {
-        return endDate;
-    }
-
-    public void setEndDate(Date endDate) {
-        this.endDate = endDate;
-    }
-
-    public String getReportDefinitionUuid() {
-        return reportDefinitionUuid;
-    }
-
-    public void setReportDefinitionUuid(String reportDefinitionUuid) {
-        this.reportDefinitionUuid = reportDefinitionUuid;
-    }
-
-    private ReportDefinitionService getReportDefinitionService() {
-        return Context.getService(ReportDefinitionService.class);
-    }
 
     public boolean isGpDhis2OrganizationUuidSet() {
         if (isBlank(syncGlobalProperties.getGlobalProperty(GP_DHIS2_ORGANIZATION_UUID))) {
@@ -186,22 +113,6 @@ public class SendReportsTask extends AbstractTask {
         return true;
     }
 
-    public String readOutputFile(String strOutput) throws Exception {
-        FileInputStream fstreamItem = new FileInputStream(OpenmrsUtil.getApplicationDataDirectory() + "sendReports");
-        DataInputStream inItem = new DataInputStream(fstreamItem);
-        BufferedReader brItem = new BufferedReader(new InputStreamReader(inItem));
-        String phraseItem;
 
-        if (!(phraseItem = brItem.readLine()).isEmpty()) {
-            strOutput = strOutput + phraseItem + System.lineSeparator();
-            while ((phraseItem = brItem.readLine()) != null) {
-                strOutput = strOutput + phraseItem + System.lineSeparator();
-            }
-        }
-
-        fstreamItem.close();
-
-        return strOutput;
-    }
 
 }
