+++ omod/pom.xml
@@ -49,6 +49,10 @@
             <artifactId>appui-omod</artifactId>
             <scope>provided</scope>
         </dependency>
+        <dependency>
+            <groupId>org.openmrs.module</groupId>
+            <artifactId>calculation-api</artifactId>
+        </dependency>
     </dependencies>
 
     <build>
