+++ omod/src/main/java/org/openmrs/module/ugandaemrsync/fragment/controller/SendReportsFragmentController.java
@@ -1,0 +1,199 @@
+/**
+ * This Source Code Form is subject to the terms of the Mozilla Public License,
+ * v. 2.0. If a copy of the MPL was not distributed with this file, You can
+ * obtain one at http://mozilla.org/MPL/2.0/. OpenMRS is also distributed under
+ * the terms of the Healthcare Disclaimer located at http://openmrs.org/license.
+ * <p>
+ * Copyright (C) OpenMRS Inc. OpenMRS is a registered trademark and the OpenMRS
+ * graphic logo is a trademark of OpenMRS Inc.
+ */
+package org.openmrs.module.ugandaemrsync.fragment.controller;
+
+import org.openmrs.api.context.Context;
+import org.openmrs.module.reporting.evaluation.EvaluationContext;
+import org.openmrs.module.reporting.evaluation.parameter.Mapped;
+import org.openmrs.module.reporting.report.ReportData;
+import org.openmrs.module.reporting.report.ReportDesign;
+import org.openmrs.module.reporting.report.ReportRequest;
+import org.openmrs.module.reporting.report.definition.ReportDefinition;
+import org.openmrs.module.reporting.report.definition.service.ReportDefinitionService;
+import org.openmrs.module.reporting.report.renderer.RenderingMode;
+import org.openmrs.module.reporting.report.renderer.TextTemplateRenderer;
+import org.openmrs.module.reporting.report.service.ReportService;
+import org.openmrs.module.ugandaemrsync.server.SyncGlobalProperties;
+import org.openmrs.module.ugandaemrsync.tasks.SendReportsTask;
+import org.openmrs.ui.framework.SimpleObject;
+import org.openmrs.ui.framework.annotation.SpringBean;
+import org.openmrs.ui.framework.fragment.FragmentModel;
+import org.openmrs.ui.framework.page.PageModel;
+import org.openmrs.util.OpenmrsUtil;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.*;
+
+import javax.servlet.http.HttpServletRequest;
+import java.io.*;
+import java.text.DateFormat;
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
+import java.util.*;
+
+import static org.openmrs.module.ugandaemrsync.UgandaEMRSyncConfig.GP_SEND_NEXT_GEN_REPORTS_SERVER_REPORT_UUIDS;
+import static org.openmrs.module.ugandaemrsync.UgandaEMRSyncConfig.JSON_REPORT_RENDERER_TYPE;
+
+/**
+ *  * Controller for a fragment that sends a report 
+ */
+public class SendReportsFragmentController {
+
+	@Autowired
+	SendReportsTask sendReportsTask;
+
+
+	SyncGlobalProperties syncGlobalProperties = new SyncGlobalProperties();
+	String reportUuids=  syncGlobalProperties.getGlobalProperty(GP_SEND_NEXT_GEN_REPORTS_SERVER_REPORT_UUIDS);
+	List<ReportDefinition> rds = getReportDefinitions(reportUuids);
+
+
+
+	public void controller(@SpringBean PageModel pageModel, @RequestParam(value = "breadcrumbOverride",
+			required = false) String breadcrumbOverride) {
+
+		pageModel.put("breadcrumbOverride", breadcrumbOverride);
+		pageModel.put("previewBody",null);
+		pageModel.put("reportDefinitions", rds);
+		pageModel.put("errorMessage", "");
+		pageModel.put("report_title",null);
+
+	}
+
+	public void post(@SpringBean PageModel pageModel,
+					 @RequestParam(value = "breadcrumbOverride", required = false) String breadcrumbOverride,
+					 @RequestParam("startDate") String periodStartDate,
+					 @RequestParam("endDate") String periodEndDate,
+					 @RequestParam("reportDefinition") String uuid) {
+
+		DateFormat dateFormat = new SimpleDateFormat("yyyy-MM-dd");
+
+		try {
+			Date startDate = dateFormat.parse(periodStartDate);
+			Date endDate = dateFormat.parse(periodEndDate);
+			String bodyText="";
+			bodyText = generateReport(uuid,startDate,endDate);
+			if(bodyText!=""){
+				String displayTitle= getReportDefinitionService().getDefinitionByUuid(uuid).getName()+" For Period \n"+
+						periodStartDate +"To " +periodEndDate;
+				pageModel.put("previewBody",bodyText);
+				pageModel.put("errorMessage", "");
+				pageModel.put("report_title",displayTitle);
+
+			}
+
+		}catch (ParseException e){
+			pageModel.put("errorMessage", e.getMessage());
+			pageModel.put("previewBody",null);
+			pageModel.put("report_title",null);
+		}
+		pageModel.put("breadcrumbOverride", breadcrumbOverride);
+		pageModel.put("reportDefinitions", rds);
+	}
+
+
+	public SimpleObject sendData(HttpServletRequest request,
+								 @RequestParam("body") String body){
+
+		String response="";
+		String status="";
+		String jsonData= body;
+
+		if(jsonData!=null){
+
+			sendReportsTask= new SendReportsTask(jsonData);
+			sendReportsTask.execute();
+			if(sendReportsTask.isSent()){
+				response= "Report successfully sent";
+				status="success";
+			}else{
+				response= "Report not sent";
+				status="failure";
+			}
+		}else{
+			response = "No Available Data to send";
+			status="failure";
+		}
+		return SimpleObject.create("status", status, "message", response);
+	}
+
+	private ReportDefinitionService getReportDefinitionService() {
+		return Context.getService(ReportDefinitionService.class);
+	}
+
+	private List<ReportDefinition> getReportDefinitions(String uuids){
+		List<String> ids= new ArrayList<>();
+		List<ReportDefinition> reportDefinitions = new ArrayList<>();
+		if(!uuids.isEmpty()&& uuids!=""){
+			ids = Arrays.asList(uuids.split(","));
+		}
+		if(ids !=null){
+			for (String id:ids) {
+				reportDefinitions.add(getReportDefinitionService().getDefinitionByUuid(id));
+			}
+		}
+		return reportDefinitions;
+	}
+
+	private String generateReport(String uuid, Date startDate, Date endDate){
+		String strOutput="";
+		try{
+			ReportDefinition rd= getReportDefinitionService().getDefinitionByUuid(uuid);
+			if (rd == null) {
+				throw new IllegalArgumentException("unable to find  report with uuid "
+						+ uuid);
+			}else {
+				List<ReportDesign> reportDesigns = Context.getService(ReportService.class).getReportDesigns(rd, TextTemplateRenderer.class, false);
+				ReportDesign reportDesign = reportDesigns.stream().filter(p -> "JSON".equals(p.getName())).findAny().orElse(null);
+				String reportRendergingMode = JSON_REPORT_RENDERER_TYPE + "!" + reportDesign.getUuid();
+				RenderingMode renderingMode = new RenderingMode(reportRendergingMode);
+				if (!renderingMode.getRenderer().canRender(rd)) {
+					throw new IllegalArgumentException("Unable to render Report with " + reportRendergingMode);
+				}
+				Map<String, Object> parameterValues = new HashMap<String, Object>();
+
+				parameterValues.put("endDate", endDate);
+				parameterValues.put("startDate", startDate);
+				EvaluationContext context = new EvaluationContext();
+				context.setParameterValues(parameterValues);
+				ReportData reportData = getReportDefinitionService().evaluate(rd, context);
+				ReportRequest reportRequest = new ReportRequest();
+				reportRequest.setReportDefinition(new Mapped<ReportDefinition>(rd, context.getParameterValues()));
+				reportRequest.setRenderingMode(renderingMode);
+				File file = new File(OpenmrsUtil.getApplicationDataDirectory() + "sendReports");
+				FileOutputStream fileOutputStream = new FileOutputStream(file);
+				renderingMode.getRenderer().render(reportData, renderingMode.getArgument(), fileOutputStream);
+
+				strOutput = readOutputFile(strOutput);
+			}
+		}catch (Exception e){
+			e.printStackTrace();
+		}
+		return strOutput;
+	}
+
+	public String readOutputFile(String strOutput) throws Exception {
+		FileInputStream fstreamItem = new FileInputStream(OpenmrsUtil.getApplicationDataDirectory() + "sendReports");
+		DataInputStream inItem = new DataInputStream(fstreamItem);
+		BufferedReader brItem = new BufferedReader(new InputStreamReader(inItem));
+		String phraseItem;
+
+		if (!(phraseItem = brItem.readLine()).isEmpty()) {
+			strOutput = strOutput + phraseItem + System.lineSeparator();
+			while ((phraseItem = brItem.readLine()) != null) {
+				strOutput = strOutput + phraseItem + System.lineSeparator();
+			}
+		}
+
+		fstreamItem.close();
+
+		return strOutput;
+	}
+	
+}
